Bigram run of consensus clustering
still need to sort out reclustering, however here we have the cluster sort through all nodes and takes a consensus on the nodes. currently if there is no agreement we are defaulting on the kmeans labeling. Thus our bigram results are the same across kmeans and the consensus clustering model so far.